requirements.txt
openai
python-dotenv
ruamel.yaml